test: update service test

+++ src/main/java/br/com/lfmelo/services/impl/BookServiceImpl.java
@@ -45,6 +45,9 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void delete(Book book) {
+        if(book == null || book.getId() == null) {
+            throw new BusinessException("Book cannot be null");
+        }
         repository.delete(book);
     }
